qa-portal-server-main/requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
flask
flasgger
paramiko
flask-cors
python-dotenv
gunicorn
pymongo

### 정경수 추가 ### 
oracledb==3.3.0
google-generativeai==0.3.2
numpy==2.3.2
requests
beautifulsoup4
selenium
webdriver-manager
requests==2.32.5
beautifulsoup4==4.12.3
psycopg2-binary==2.9.9
pywinrm